caustaza/requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.3.0
click==8.1.3
Django==4.2.2
django-ckeditor==6.5.1
django-environ==0.10.0
django-filter==23.2
django-jazzmin==2.6.0
django-js-asset==2.0.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
inflection==0.5.1
jsonschema==4.17.3
Markdown==3.4.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.7.0
psycopg==3.1.9
psycopg2==2.9.6
pyrsistent==0.19.3
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
